Use arrays instead of strings in pointLocation
Writing strings and then parsing them will results in a vastly
bigger time consumption than just providing the values as they are.

This changes the pointLocation function to use arrays of the form
array(lng,lat). An argument with string coordinates is still
accepted, but converted. Omitting the 'x' and 'y' keys in favor
of integer might speed things up additionally.

Requires the table to store the coordinates as DOUBLE!

Signed-off-by: Adrian Schmutzler <adrian.schmutzler@uni-bayreuth.de>