Implement two polyhood tables and add shortcut for extreme values
This patch implements the two tables created in the previous patch.

Exploiting the new structure, we can redesign the polyhood check
to check whether the point is within a square defined by the
extreme values of the coordinates, i.e. the latitude is smaller
than the one of the southmost vertex etc.

Although this is an extra check, it is much faster if we can skip
a lot of full polygon checks. Since we expect mostly "normal"
hoods and only few routers in polyhoods, this should be the case.

Signed-off-by: Adrian Schmutzler <adrian.schmutzler@uni-bayreuth.de>