Don't store first polygon vertex twice
So far, the first and last point of every polygon in the database
has been the same.

This is unnecessary, so this patch changes behavior to make stored
data points unique. As a result, we need to add the missing final
point in the interpretation of the data. This has the advantage
that errors do to typos or rounding problems will be eliminated.

The general concept is as follows:

DATA SOURCES (database, hoods.php, etc.) should contain only unique
points.

DATA INTERPRETERS (Monitoring, map.html, etc.) will then add the
final point and display the result correctly.

Signed-off-by: Adrian Schmutzler <adrian.schmutzler@uni-bayreuth.de>